requirements.txt
fastapi==0.115.0
psycopg2-binary==2.9.10
python-dotenv==1.2.2
uvicorn==0.30.0